Snapshot of develop 2018-05-13 (Internal repo commit 4c15cd2)
* 2019-05-09 (4c15cd2) Move LaTeX files to latest formatting style from main repo
* 2019-05-06 (c58759f) Squashed commit of the following:
* 2019-05-01 (d47f696) Merge remote-tracking branch 'origin/documentation-v1.5.0' into v1.5.0
* 2019-04-30 (46600f9) removing now-known-bad-practice non-default value for messageSize property of file_read (AV-5465)
* 2019-04-30 (4e916f5) removing now-known-bad-practice non-default value for messageSize property of file_read (AV-5465)
*   2019-04-22 (d0eb1bc) Merge branch 'AV-5429' into develop
|\
| * 2019-04-22 (ad414ca) AV-5429, which is also a bugfix for AV-5423
* | 2019-04-19 (44d8e56) Merge branch 'AV-5427' into develop
* | 2019-04-19 (7fb5ff9) AV-5427
|/
* 2019-04-12 (ee6826a) Merge pull request #6 in AV/av.bsp.e310 from bugfix-library-mk to v1.5.0

+++ applications/FSK/scripts/gen_rrcos_taps.py
@@ -72,7 +72,7 @@ def main():
 
     fo = open(sys.argv[6], 'w')
     #print "\tName of the output file:", fo.name
-    for i in taps[0:np.ceil(length/2.0)]:
+    for i in taps[0:int(np.ceil(length/2.0))]:
         stringy = ''.join(str(i)+'\n')
         fo.write(stringy)
     fo.close()
